Refactor event and entity structures to align with the new DTO-based design and improve modularity.

+++ src/main/java/fr/ses10doigts/tradeIO5/model/dto/tree/decision/ActionStep.java
@@ -2,10 +2,12 @@ package fr.ses10doigts.tradeIO5.model.dto.tree.decision;
 
 import fr.ses10doigts.tradeIO5.model.enumerate.tree.decision.ExecutionAction;
 
+import java.math.BigDecimal;
+
 
 public record ActionStep(
         String stepId,
         ExecutionAction executionAction,
-        double quantity) {
-
+        BigDecimal quantity
+) {
 }
